# monolithic/backend/app/admin/service/cmd/server/assets/openapi.yaml
...
openapi: 3.0.3
info:
    title: Kratos巨石应用实践API
    description: Kratos巨石应用实践API
    contact:
        name: tx7do
        url: https://github.com/tx7do/kratos-monolithic-demo
        email: yanglinbo@gmail.com
    license:
        name: MIT License
        url: https://github.com/tx7do/kratos-monolithic-demo/blob/master/LICENSE
    version: "1.0"
paths:
...
    /admin/v1/positions:
        get:
            tags:
                - PositionService
            description: 查询职位列表
            operationId: PositionService_ListPosition
            parameters:
                - name: page
                  in: query
                  description: 当前页码
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  description: 每页的行数
                  schema:
                    type: integer
                    format: int32
                - name: query
                  in: query
                  description: 与过滤参数
                  schema:
                    type: string
                - name: or
                  in: query
                  description: 或过滤参数
                  schema:
                    type: string
                - name: orderBy
                  in: query
                  description: 排序条件
                  schema:
                    type: array
                    items:
                        type: string
                - name: nopaging
                  in: query
                  description: 是否不分页
                  schema:
                    type: boolean
                - name: fieldMask
                  in: query
                  description: 字段掩码
                  schema:
                    type: string
                    format: field-mask
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListPositionResponse'
            security:
                - OAuth2PasswordBearer: []
        post:
            tags:
                - PositionService
            description: 创建职位
            operationId: PositionService_CreatePosition
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreatePositionRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
            security:
                - OAuth2PasswordBearer: []
    /admin/v1/positions/{id}:
        get:
            tags:
                - PositionService
            description: 查询职位详情
            operationId: PositionService_GetPosition
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Position'
            security:
                - OAuth2PasswordBearer: []
        delete:
            tags:
                - PositionService
            description: 删除职位
            operationId: PositionService_DeletePosition
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
                - name: operatorId
                  in: query
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content: {}
            security:
                - OAuth2PasswordBearer: []
    /admin/v1/positions/{position.id}:
        put:
            tags:
                - PositionService
            description: 更新职位
            operationId: PositionService_UpdatePosition
            parameters:
                - name: position.id
                  in: path
                  required: true
                  schema:
                    type: string
                - name: operatorId
                  in: query
                  schema:
                    type: integer
                    format: uint32
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Position'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
            security:
                - OAuth2PasswordBearer: []
...
components:
    schemas:
...
        CreatePositionRequest:
            type: object
            properties:
                position:
                    $ref: '#/components/schemas/Position'
                operatorId:
                    type: integer
                    format: uint32
            description: 创建职位 - 请求
...
        ListPositionResponse:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/Position'
                total:
                    type: integer
                    format: int32
            description: 获取职位列表 - 答复
...
        Position:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
                name:
                    type: string
                parentId:
                    type: integer
                    format: uint32
                orderNo:
                    type: integer
                    format: int32
                code:
                    type: string
                status:
                    enum:
                        - ON
                        - OFF
                    type: string
                    default: ON
                    description: 状态
                remark:
                    type: string
                createTime:
                    type: string
                updateTime:
                    type: string
                deleteTime:
                    type: string
            description: 职位
...
    securitySchemes:
        OAuth2PasswordBearer:
            type: oauth2
            flows:
                password:
                    tokenUrl: /admin/v1/login
                    refreshUrl: /admin/v1/refresh_token
                    scopes: {}
...